requirements.txt
pandas
sentence-transformers
faiss-cpu
pillow
pytesseract
easyocr